Day 3_Part 4A
This part mainly focuses on extending the Ride class with the ability to store information about visitors who have ridden the ride. Specifically, we introduce a LinkedList collection to store visitors (of type Visitor) and add a series of methods to interact with the collection, including adding visitors to the ride history, checking whether a visitor is in the history, getting the number of visitors in the history, and printing the details of all the visitors who have ridden the ride, and each of the methods has a corresponding prompt message. At the same time, it is stipulated that the Ride object is created in the partFourA method of AssignmentTwo class and these methods are used to demonstrate the functions, such as adding at least 5 tourists, checking the specific tourists, printing the number of tourists and all the information of tourists, etc., so as to improve the functions and validate the operation procedures of the Ride class for the management of the tourists' ride history.