Day 3-Part 4B
I've added important functionality to the code by adding the sortVisitors method to the RideInterface interface for sorting visitor information. Implement this method in the Ride class, use Collections.sort and VisitorComparator to customise the comparator to sort by age and first time visitor rule and output hints. Meanwhile, we improve the removeVisitorFromQueue method to improve the flow of visitor data. Also write VisitorComparator class to define the comparison rules. Demonstrate the sorting function in the partFourB method of AssignmentTwo class, add tourists, print the information before and after sorting, so as to make the processing and display of tourists' ride history data more perfect.

Translated with DeepL.com (free version)