Day 4-Part 5
This part of my code mainly implements the functionality related to running the ride in cycles. In the Ride class I defined, I added two attributes, one is the maxRider attribute, which can be used to specify how many visitors I can carry in one cycle, and it is stipulated that there must be at least 1 visitor before the ride can run; the other is the numOfCycles attribute, which is used to record the number of times the ride has run, and the default value is set to 0, and every time the ride runs, the value will be increased to 1. The other one is numOfCycles, which is used to record the number of times the ride has been run.

Then I also implemented the RunOneCycle method in the Ride class. This method has some important logic judgement, if there is no ride operator assigned to me, then I can't run it, and I will output the corresponding message; if there is no waiting tourist in my queue, I also can't run it, and I will also output the corresponding message. If all the conditions are met, I will remove the tourists from the queue according to the number of maxRider, and add them to the collection used to record the ride history.

I also demonstrated the functionality in the partFive method of the AssignmentTwo class. I create a new Ride object, add at least 10 tourists to the queue using the methods I created, and then print out the information for all the tourists in the queue before the run. Then I'll run a cycle and print out all the visitors in the queue at that point, and finally print out all the visitors in the history collection, to show the complete flow of the ride's cycle-related functionality.