Day 4-Part 6
In the `partFive` method of the `AssignmentTwo` class, I'm going to do all of these things:
First, I'll create a new `Ride` object, which will be the basis for the subsequent operations. Then, using the methods I created earlier, the first step is to add at least 10 tourists to the queue of the `Ride` object, so that I can simulate a queue of tourists waiting to ride the ride. Then, I'll print out the information of all the tourists in the queue, so that I can check and confirm the tourists I've added. After that, I will run a cycle of the ride, so that some of the tourists can go on the ride. After this cycle, I will print out the information of the remaining visitors in the queue to see which visitors are still in the queue. Finally, I'll print out all the visitors in the set of visitors who have ridden the rides, to show a full picture of the visitors at different stages of the process.

When it comes to the ‘Module 5 - Input/Output’ section, a lot of applications value the ability to back up data to a file and restore data on demand, so I'm going to add this feature to my application for this assignment. Give a tip about exporting and importing data, a common way of storing data in a file that needs to be imported later is to use the comma-separated values (CSV) format, like for an ‘animal’ object that contains information about species, breed, colour, name and age, you can put its information in one line. ‘species, breed, colour, name, age’ such a format stored in the file it, read the file can be divided by commas to get each line of each value, and then create new objects.

As for ‘Part 6 - Writing to a file’, my `Ride` class lacked the ability to back up the information of visitors who have ridden the rides before (as you can see in Part 3, Related Operations). For this part of the assignment, I have to add a method called `exportRideHistory` to the `Ride` class, and use this method to write the details of all the tourists who have been on the rides (i.e., those stored in the `LinkedList`) to a single file, with one line each for each tourists' details. At the same time, I have to make sure that I add all the appropriate exception and error messages, but I don't need to back up the visitor information in the queue (Part 2 deals with queue operations). In the `partSix` method of the `AssignmentTwo` class, I have to create a new `Ride` object, add at least 5 tourists to the `Ride` object (or to be precise, to its corresponding `LinkedList`), and then export the tourists' information to a file.