Part 7
In this section, it is first mentioned that an important point for many programs is the ability to back up data to a file and restore it as needed. For the Ride class in the program, it originally lacks the ability to back up the information of the visitors who have taken the rides and restore this visitor information.

I added the method as requested. In part six, I added the `exportRideHistory` method to the `Ride` class to write the details of the guests who have ridden the ride (stored in a `LinkedList`) to a file, one guest per line. and also added appropriate exception handling and error messages. The operation is demonstrated in the `partSix` method of the `AssignmentTwo` class. Create a new amusement ride, add at least five visitors and export the visitor information to a file.

In part seven, add the `importRideHistory` method to the `Ride` class so that it can read the file created in the previous section. When reading, add the visitor information to the `LinkedList`. Again, add appropriate exception handling and error messages. Even if adding the visitor to the file, even if the attempt to add the visitor information to the LinkedList fails, the score can still be read from the file. This is demonstrated in the partSeven method of the AssignmentTwo class by creating a new ride, importing the file, printing the number of visitors to confirm that the import quantity is correct, and printing the visitor information to confirm that the import details are accurate.